remove deploy processor, move logic to manager
