+++ src/snowflake/cli/plugins/nativeapp/manager.py
@@ -6,6 +6,7 @@ from pathlib import Path
 from textwrap import dedent
 from typing import List, Optional
 
+import jinja2
 from snowflake.cli.api.console import cli_console as cc
 from snowflake.cli.api.exceptions import SnowflakeSQLExecutionError
 from snowflake.cli.api.project.definition import (
@@ -27,12 +28,21 @@ from snowflake.cli.plugins.nativeapp.artifacts import (
     translate_artifact,
 )
 from snowflake.cli.plugins.nativeapp.constants import (
+    ALLOWED_SPECIAL_COMMENTS,
+    COMMENT_COL,
     ERROR_MESSAGE_606,
     ERROR_MESSAGE_2043,
+    INTERNAL_DISTRIBUTION,
     NAME_COL,
     OWNER_COL,
+    SPECIAL_COMMENT,
 )
-from snowflake.cli.plugins.nativeapp.exceptions import UnexpectedOwnerError
+from snowflake.cli.plugins.nativeapp.exceptions import (
+    ApplicationPackageAlreadyExistsError,
+    InvalidPackageScriptError,
+    MissingPackageScriptError,
+    UnexpectedOwnerError,
+)
 from snowflake.cli.plugins.object.stage.diff import (
     DiffResult,
     stage_diff,
@@ -327,3 +337,99 @@ class NativeAppManager(SqlExecutionMixin):
         """Returns the URL that can be used to visit this app via Snowsight."""
         name = unquote_identifier(self.app_name)
         return make_snowsight_url(self._conn, f"/#/apps/application/{name}")
+
+    def create_app_package(self) -> None:
+        """
+        Creates the application package with our up-to-date stage if none exists.
+        """
+
+        # 1. Check for existing existing application package
+        show_obj_row = self.get_existing_app_pkg_info()
+
+        if show_obj_row:
+            # 1. Check for the right owner role
+            ensure_correct_owner(
+                row=show_obj_row, role=self.package_role, obj_name=self.package_name
+            )
+
+            # 2. Check distribution of the existing application package
+            actual_distribution = self.get_app_pkg_distribution_in_snowflake
+            if not self.verify_project_distribution(actual_distribution):
+                cc.warning(
+                    f"Continuing to execute `snow app run` on application package {self.package_name} with distribution '{actual_distribution}'."
+                )
+
+            # 3. If actual_distribution is external, skip comment check
+            if actual_distribution == INTERNAL_DISTRIBUTION:
+                row_comment = show_obj_row[COMMENT_COL]
+
+                if row_comment not in ALLOWED_SPECIAL_COMMENTS:
+                    raise ApplicationPackageAlreadyExistsError(self.package_name)
+
+            return
+
+        # If no application package pre-exists, create an application package, with the specified distribution in the project definition file.
+        with self.use_role(self.package_role):
+            cc.step(f"Creating new application package {self.package_name} in account.")
+            self._execute_query(
+                dedent(
+                    f"""\
+                    create application package {self.package_name}
+                        comment = {SPECIAL_COMMENT}
+                        distribution = {self.package_distribution}
+                """
+                )
+            )
+
+    def _apply_package_scripts(self) -> None:
+        """
+        Assuming the application package exists and we are using the correct role,
+        applies all package scripts in-order to the application package.
+        """
+        env = jinja2.Environment(
+            loader=jinja2.loaders.FileSystemLoader(self.project_root),
+            keep_trailing_newline=True,
+            undefined=jinja2.StrictUndefined,
+        )
+
+        queued_queries = []
+        for relpath in self.package_scripts:
+            try:
+                template = env.get_template(relpath)
+                result = template.render(dict(package_name=self.package_name))
+                queued_queries.append(result)
+
+            except jinja2.TemplateNotFound as e:
+                raise MissingPackageScriptError(e.name)
+
+            except jinja2.TemplateSyntaxError as e:
+                raise InvalidPackageScriptError(e.name, e)
+
+            except jinja2.UndefinedError as e:
+                raise InvalidPackageScriptError(relpath, e)
+
+        # once we're sure all the templates expanded correctly, execute all of them
+        try:
+            if self.package_warehouse:
+                self._execute_query(f"use warehouse {self.package_warehouse}")
+
+            for i, queries in enumerate(queued_queries):
+                cc.step(f"Applying package script: {self.package_scripts[i]}")
+                self._execute_queries(queries)
+        except ProgrammingError as err:
+            generic_sql_error_handler(
+                err, role=self.package_role, warehouse=self.package_warehouse
+            )
+
+    def deploy(self) -> None:
+        """app deploy process"""
+
+        # 1. Create an empty application package, if none exists
+        self.create_app_package()
+
+        with self.use_role(self.package_role):
+            # 2. now that the application package exists, create shared data
+            self._apply_package_scripts()
+
+            # 3. Upload files from deploy root local folder to the above stage
+            diff = self.sync_deploy_root_with_stage(self.package_role)
