+++ tests/nativeapp/utils.py
@@ -5,12 +5,10 @@ TEARDOWN_MODULE = "snowflake.cli.plugins.nativeapp.teardown_processor"
 TYPER_CONFIRM = "typer.confirm"
 RUN_MODULE = "snowflake.cli.plugins.nativeapp.run_processor"
 VERSION_MODULE = "snowflake.cli.plugins.nativeapp.version.version_processor"
-DEPLOY_MODULE = "snowflake.cli.plugins.nativeapp.deploy_processor"
 
 TEARDOWN_PROCESSOR = f"{TEARDOWN_MODULE}.NativeAppTeardownProcessor"
 NATIVEAPP_MANAGER = f"{NATIVEAPP_MODULE}.NativeAppManager"
 RUN_PROCESSOR = f"{RUN_MODULE}.NativeAppRunProcessor"
-DEPLOY_PROCESSOR = f"{DEPLOY_MODULE}.NativeAppDeployProcessor"
 
 NATIVEAPP_MANAGER_EXECUTE = f"{NATIVEAPP_MANAGER}._execute_query"
 NATIVEAPP_MANAGER_EXECUTE_QUERIES = f"{NATIVEAPP_MANAGER}._execute_queries"
@@ -19,6 +17,9 @@ NATIVEAPP_MANAGER_APP_PKG_DISTRIBUTION_IN_SF = (
 )
 NATIVEAPP_MANAGER_IS_APP_PKG_DISTRIBUTION_SAME = (
     f"{NATIVEAPP_MANAGER}.verify_project_distribution"
+)
+NATIVEAPP_MANAGER_GET_EXISTING_APP_PKG_INFO = (
+    f"{NATIVEAPP_MANAGER}.get_existing_app_pkg_info"
 )
 
 TEARDOWN_PROCESSOR_GET_EXISTING_APP_INFO = f"{TEARDOWN_PROCESSOR}.get_existing_app_info"
@@ -29,10 +30,6 @@ TEARDOWN_PROCESSOR_IS_CORRECT_OWNER = f"{TEARDOWN_MODULE}.ensure_correct_owner"
 TEARDOWN_PROCESSOR_DROP_GENERIC_OBJECT = f"{TEARDOWN_PROCESSOR}.drop_generic_object"
 
 RUN_PROCESSOR_GET_EXISTING_APP_INFO = f"{RUN_PROCESSOR}.get_existing_app_info"
-
-DEPLOY_PROCESSOR_GET_EXISTING_APP_PKG_INFO = (
-    f"{DEPLOY_PROCESSOR}.get_existing_app_pkg_info"
-)
 
 FIND_VERSION_FROM_MANIFEST = f"{VERSION_MODULE}.find_version_info_in_manifest_file"
